implementing logic to let declare complex cache requirements that depends on dynamic factors

+++ user-service/src/main/java/com/example/user/infrastructure/cache/CacheRequirements.java
@@ -18,6 +18,6 @@ public class CacheRequirements {
     @Setter
     public static class Requirement {
         private String cacheName;
-        private int ttlMinutes;
+        private Long ttlMinutes;
     }
 }
